fix bug

+++ app/crawler/crawler_module/controller/company.py
@@ -3,11 +3,17 @@ from app.crawler.crawler_module.utils.utils import generate_company_file
 from app.crawler.crawler_module.utils.file import delete_file
 from app.crawler.crawler_module.parser.company import parse_company
 from app.crawler.crawler_module.crawler.company import collect_company
-from app.crawler.db.company.insert import enrich_company
+from app.crawler.db.company.insert import enrich_company, check_company_exist
 
 
 def handle_company_consumer(event: CompanyEvent):
+  if (check_company_exist(event.company_id)):
+    return
   collect_company(generate_company_file(event.company_id), event.url)
-  company = parse_company(event.company_id, generate_company_file(event.company_id))
-  enrich_company(company)
-  delete_file(generate_company_file(event.company_id))
+  try:
+    company = parse_company(event.company_id, generate_company_file(event.company_id))
+    enrich_company(company)
+  except Exception as err:
+    print(f'Error getting company {event.company_id} {err}')
+  finally:
+    delete_file(generate_company_file(event.company_id))
